Updated this so that most of it is working
I wasn't sure how SourceToString was working so I commented it out, but it's still there if it needs to be implemented.

We need to find the specific node types that declarations and references are under. There's a great post about the differences between the two in the discussions section of D2L. Tomorrow I'll finish this off and add a couple of other things.